Move ImageActivity UI classes to separate packages
